Use double quotes, not single quotes.
Douglas Crockfort:

When I first met JavaScript, it was surprised that it had the two kinds
of quotes, and I tried to make sense of it, using single for internal
text, and double for external.

But eventually I realized that distinction isn't worth the clutter and
confusion that comes from having two when only one is needed. I decided
to go with double because that is what JSON uses, and it avoids errors
caused by the overloading of apostrophe. I have been bitten by that.

In general, I am looking for ways to make the language smaller and
better. Quotes fall in the same class as null & undefined. We really
don't need both.

I tried it out on some of my own code, and I think it is an improvement.
Eventually, I may add option.single to JSLint.

https://plus.google.com/+%D0%95%D0%B2%D0%B3%D0%B5%D0%BD%D0%B8%D0%B9%D0%9E%D1%80%D0%B5%D1%85%D0%BE%D0%B2/posts/dykmroX6oYH